requirements.txt
# Testing
pytest==9.0.1

# HTTP client
requests==2.32.5

# Allure
allure-pytest==2.13.2
